move more tests to apollo-runtime
